Add type attribute validation
